Update egg_bind and js_bind so that they should properly mirror eggdrop's bind function.
Fixed several memory leaks throughout the application.  Added cleanup code for the Javascript api's to free memory used 
by Spidermonkey.  The only change I'm not sure about is the addition of JS_DestroyScript in js_eval_file.  I'm not sure if 
destroying the script at that point is proper, or if it should not be done.



git-svn-id: https://trollbot.svn.sourceforge.net/svnroot/trollbot/trunk@62 79ca4b1d-1696-da11-8a5f-000b6a37957f